requirements.txt
fastapi[all]
google-auth
google-auth-oauthlib
uvicorn
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
fastapi
pydantic
